Adds Solarized theme for vim (but doesn't enable)
Solarized is a highly popular color scheme that comes in both light and
dark variants.

Ideally, we'd be using the 16-color scheme for Solarized instead of the
approximated 256-color version, but automating that configuration is
tricky, and as it stands, would break too many other things.

As a bonus, Powerline includes a Solarized theme, which we now use. It
fits in well with both Solarized and Jellybeans.