Make vim more responsive when esc is pressed
Vim waits a certain amount of time after esc is pressed before switching
to normal mode in case a special key like the arrow keys, or F1, F2, etc
is pressed (as they are passed to vim escaped with the esc key). This
reduces the timeout to 1/10th of a second, which should still be slow
enough for most machines, but is fast enough that it is hard to notice.

Another alternative solution to this problem would be to disable
esc-escaping, but that would also disable the function and arrow keys.