Adds MarcWeber/vim-addon-local-vimrc
I was using `set exrc` for loading project-level vimrc files, but it
doesn't traverse up the directory tree. Instead it only looks in the
current working directory. This will the traverse the whole tree.

I'm also asked before a vimrc file is sourced (to be safe), and the
results of my approval/denial is stored via the file hash.