[zsh] Expand sudo arguments, add `sudo -E` alias
Expanding sudo's arguments allows me to use my other aliases with sudo.
The new alias, `sd` passes the environment through, allowing
configuration files and the like to be passed through to the other side.

N.B. the -E has it's own set of issues, some security related, so we
aren't applying it to the standard `sudo` command by default.