[terminfo] Rename *-256color-italic to *-256color
This overwrites the existing terminfo entry, which is hacky, and not how
it's supposed to work, but far too many programs seem to inspect $TERM,
and freak out when the value isn't in their database. Plus, all sorts of
issues are created when $TERM is passed through to ssh, mosh, sudo, etc.

So overwriting the normal *-256color files seems to actually be the
least hacky approach.